Add a button to send individual profiles to CIS.
* Move admin forms in a separate file.
